requirements.txt
autopep8==1.7.0
channels==4.0.0
channels-redis==4.0.0
Django==4.1.2
django-active-link==0.1.8
django-appconf==1.0.5
django-bootstrap-modal-forms==2.2.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cleanup==6.0.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-filter==22.1
django-formtools==2.4
django-reset-migrations==0.4.0
django-stronghold==0.4.0
django-tenant-users==1.1.0
django-tenants==3.4.5
django-tenants-celery-beat==0.2.0
django-timezone-field==5.0
django-tus==0.5.0
django-view-breadcrumbs==2.4.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
feedparser==6.0.10
Pillow==9.2.0
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pylint==2.15.4
python-decouple==3.6
redis==4.3.4
requests==2.28.1
tzdata==2022.5
whitenoise==6.2.0